# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Galaxy ERP",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/impuestos/libro-ventas": {
      "get": {
        "tags": [
          "Impuestos y Libros Fiscales SENIAT"
        ],
        "summary": "Consultar Libro Ventas",
        "operationId": "consultar_libro_ventas_api_v1_impuestos_libro_ventas_get",
        "parameters": [
          {
            "name": "anio",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2050,
              "minimum": 2020,
              "title": "Anio"
            }
          },
          {
            "name": "mes",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 12,
              "minimum": 1,
              "title": "Mes"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibroVentasResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/impuestos/libro-ventas/excel": {
      "get": {
        "tags": [
          "Impuestos y Libros Fiscales SENIAT"
        ],
        "summary": "Descargar Libro Ventas Excel",
        "operationId": "descargar_libro_ventas_excel_api_v1_impuestos_libro_ventas_excel_get",
        "parameters": [
          {
            "name": "anio",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2050,
              "minimum": 2020,
              "title": "Anio"
            }
          },
          {
            "name": "mes",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 12,
              "minimum": 1,
              "title": "Mes"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/impuestos/libro-compras": {
      "get": {
        "tags": [
          "Impuestos y Libros Fiscales SENIAT"
        ],
        "summary": "Consultar Libro Compras",
        "operationId": "consultar_libro_compras_api_v1_impuestos_libro_compras_get",
        "parameters": [
          {
            "name": "anio",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2050,
              "minimum": 2020,
              "title": "Anio"
            }
          },
          {
            "name": "mes",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 12,
              "minimum": 1,
              "title": "Mes"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibroComprasResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/impuestos/libro-compras/excel": {
      "get": {
        "tags": [
          "Impuestos y Libros Fiscales SENIAT"
        ],
        "summary": "Descargar Libro Compras Excel",
        "operationId": "descargar_libro_compras_excel_api_v1_impuestos_libro_compras_excel_get",
        "parameters": [
          {
            "name": "anio",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 2050,
              "minimum": 2020,
              "title": "Anio"
            }
          },
          {
            "name": "mes",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 12,
              "minimum": 1,
              "title": "Mes"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/impuestos/retenciones-iva": {
      "get": {
        "tags": [
          "Impuestos y Libros Fiscales SENIAT"
        ],
        "summary": "Listar Retenciones Iva",
        "operationId": "listar_retenciones_iva_api_v1_impuestos_retenciones_iva_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComprobanteIvaResponse"
                  },
                  "title": "Response Listar Retenciones Iva Api V1 Impuestos Retenciones Iva Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/impuestos/retenciones-iva/{comprobante_id}/pdf": {
      "get": {
        "tags": [
          "Impuestos y Libros Fiscales SENIAT"
        ],
        "summary": "Descargar Retencion Iva Pdf",
        "operationId": "descargar_retencion_iva_pdf_api_v1_impuestos_retenciones_iva__comprobante_id__pdf_get",
        "parameters": [
          {
            "name": "comprobante_id",
            "in": "path",
            "required": true,
            "schema": {
              "title": "Comprobante Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/dispositivos": {
      "post": {
        "tags": [
          "Sincronizaci\u00f3n Offline y Preventa"
        ],
        "summary": "Registrar Dispositivo",
        "operationId": "registrar_dispositivo_api_v1_sync_dispositivos_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DispositivoSyncCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispositivoSyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/catalogos": {
      "get": {
        "tags": [
          "Sincronizaci\u00f3n Offline y Preventa"
        ],
        "summary": "Descargar Catalogos",
        "operationId": "descargar_catalogos_api_v1_sync_catalogos_get",
        "parameters": [
          {
            "name": "desde",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Desde"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogosSyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/bloques": {
      "post": {
        "tags": [
          "Sincronizaci\u00f3n Offline y Preventa"
        ],
        "summary": "Solicitar Bloque Correlativo",
        "operationId": "solicitar_bloque_correlativo_api_v1_sync_bloques_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BloqueCorrelativoRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BloqueCorrelativoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/operaciones": {
      "post": {
        "tags": [
          "Sincronizaci\u00f3n Offline y Preventa"
        ],
        "summary": "Procesar Operaciones Offline",
        "operationId": "procesar_operaciones_offline_api_v1_sync_operaciones_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchOperacionesSyncRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchOperacionesSyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchOperacionesSyncRequest": {
        "properties": {
          "dispositivo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dispositivo Id"
          },
          "operaciones": {
            "items": {
              "$ref": "#/components/schemas/OperacionSyncItem"
            },
            "type": "array",
            "title": "Operaciones"
          }
        },
        "type": "object",
        "required": [
          "dispositivo_id",
          "operaciones"
        ],
        "title": "BatchOperacionesSyncRequest"
      },
      "BatchOperacionesSyncResponse": {
        "properties": {
          "resultados": {
            "items": {
              "$ref": "#/components/schemas/OperacionSyncResultado"
            },
            "type": "array",
            "title": "Resultados"
          }
        },
        "type": "object",
        "required": [
          "resultados"
        ],
        "title": "BatchOperacionesSyncResponse"
      },
      "BloqueCorrelativoRequest": {
        "properties": {
          "dispositivo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dispositivo Id"
          },
          "tipo_documento": {
            "type": "string",
            "maxLength": 30,
            "title": "Tipo Documento",
            "default": "PRESUPUESTO"
          },
          "serie": {
            "type": "string",
            "maxLength": 10,
            "title": "Serie",
            "default": ""
          },
          "cantidad": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Cantidad",
            "default": 50
          }
        },
        "type": "object",
        "required": [
          "dispositivo_id"
        ],
        "title": "BloqueCorrelativoRequest"
      },
      "BloqueCorrelativoResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "dispositivo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dispositivo Id"
          },
          "tipo_documento": {
            "type": "string",
            "title": "Tipo Documento"
          },
          "serie": {
            "type": "string",
            "title": "Serie"
          },
          "desde_numero": {
            "type": "integer",
            "title": "Desde Numero"
          },
          "hasta_numero": {
            "type": "integer",
            "title": "Hasta Numero"
          },
          "ultimo_usado": {
            "type": "integer",
            "title": "Ultimo Usado"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dispositivo_id",
          "tipo_documento",
          "serie",
          "desde_numero",
          "hasta_numero",
          "ultimo_usado"
        ],
        "title": "BloqueCorrelativoResponse"
      },
...
      "CatalogosSyncResponse": {
        "properties": {
          "sync_token": {
            "type": "integer",
            "title": "Sync Token"
          },
          "productos": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Productos"
          },
          "clientes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Clientes"
          },
          "almacenes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Almacenes"
          },
          "existencias": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Existencias"
          },
          "alicuotas_iva": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Alicuotas Iva"
          },
          "tasa_actual_usd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tasa Actual Usd"
          }
        },
        "type": "object",
        "required": [
          "sync_token",
          "productos",
          "clientes",
          "almacenes",
          "existencias",
          "alicuotas_iva",
          "tasa_actual_usd"
        ],
        "title": "CatalogosSyncResponse"
      },
...
      "ComprobanteIvaResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "proveedor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Proveedor Id"
          },
          "compra_id": {
            "type": "string",
            "format": "uuid",
            "title": "Compra Id"
          },
          "numero_comprobante": {
            "type": "string",
            "title": "Numero Comprobante"
          },
          "periodo_fiscal": {
            "type": "string",
            "title": "Periodo Fiscal"
          },
          "fecha_emision": {
            "type": "string",
            "format": "date",
            "title": "Fecha Emision"
          },
          "base_imponible_ves": {
            "type": "string",
            "title": "Base Imponible Ves"
          },
          "monto_iva_ves": {
            "type": "string",
            "title": "Monto Iva Ves"
          },
          "porcentaje_retencion": {
            "type": "string",
            "title": "Porcentaje Retencion"
          },
          "monto_retenido_ves": {
            "type": "string",
            "title": "Monto Retenido Ves"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_id",
          "proveedor_id",
          "compra_id",
          "numero_comprobante",
          "periodo_fiscal",
          "fecha_emision",
          "base_imponible_ves",
          "monto_iva_ves",
          "porcentaje_retencion",
          "monto_retenido_ves",
          "created_at"
        ],
        "title": "ComprobanteIvaResponse"
      },
...
      "DispositivoSyncCreate": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 100,
            "title": "Nombre"
          },
          "identificador_unico": {
            "type": "string",
            "maxLength": 120,
            "title": "Identificador Unico"
          },
          "usuario_id": {
            "type": "string",
            "format": "uuid",
            "title": "Usuario Id"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "identificador_unico",
          "usuario_id"
        ],
        "title": "DispositivoSyncCreate"
      },
      "DispositivoSyncResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "identificador_unico": {
            "type": "string",
            "title": "Identificador Unico"
          },
          "usuario_id": {
            "type": "string",
            "format": "uuid",
            "title": "Usuario Id"
          },
          "ultimo_token": {
            "type": "integer",
            "title": "Ultimo Token"
          },
          "activo": {
            "type": "boolean",
            "title": "Activo"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "identificador_unico",
          "usuario_id",
          "ultimo_token",
          "activo",
          "created_at"
        ],
        "title": "DispositivoSyncResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LibroComprasResponse": {
        "properties": {
          "periodo_fiscal": {
            "type": "string",
            "title": "Periodo Fiscal"
          },
          "filas": {
            "items": {
              "$ref": "#/components/schemas/LibroComprasFila"
            },
            "type": "array",
            "title": "Filas"
          },
          "total_compras_ves": {
            "type": "string",
            "title": "Total Compras Ves"
          },
          "total_exento_ves": {
            "type": "string",
            "title": "Total Exento Ves"
          },
          "total_base_ves": {
            "type": "string",
            "title": "Total Base Ves"
          },
          "total_iva_ves": {
            "type": "string",
            "title": "Total Iva Ves"
          },
          "total_iva_retenido_ves": {
            "type": "string",
            "title": "Total Iva Retenido Ves"
          }
        },
        "type": "object",
        "required": [
          "periodo_fiscal",
          "filas",
          "total_compras_ves",
          "total_exento_ves",
          "total_base_ves",
          "total_iva_ves",
          "total_iva_retenido_ves"
        ],
        "title": "LibroComprasResponse"
      },
      "LibroVentasFila": {
        "properties": {
          "operacion": {
            "type": "integer",
            "title": "Operacion"
          },
          "fecha": {
            "type": "string",
            "format": "date",
            "title": "Fecha"
          },
          "rif_cliente": {
            "type": "string",
            "title": "Rif Cliente"
          },
          "nombre_cliente": {
            "type": "string",
            "title": "Nombre Cliente"
          },
          "numero_factura": {
            "type": "string",
            "title": "Numero Factura"
          },
          "numero_control": {
            "type": "string",
            "title": "Numero Control"
          },
          "total_ventas_ves": {
            "type": "string",
            "title": "Total Ventas Ves"
          },
          "ventas_exentas_ves": {
            "type": "string",
            "title": "Ventas Exentas Ves"
          },
          "base_imponible_ves": {
            "type": "string",
            "title": "Base Imponible Ves"
          },
          "iva_ves": {
            "type": "string",
            "title": "Iva Ves"
          },
          "iva_retenido_ves": {
            "type": "string",
            "title": "Iva Retenido Ves"
          }
        },
        "type": "object",
        "required": [
          "operacion",
          "fecha",
          "rif_cliente",
          "nombre_cliente",
          "numero_factura",
          "numero_control",
          "total_ventas_ves",
          "ventas_exentas_ves",
          "base_imponible_ves",
          "iva_ves",
          "iva_retenido_ves"
        ],
        "title": "LibroVentasFila"
      },
      "LibroVentasResponse": {
        "properties": {
          "periodo_fiscal": {
            "type": "string",
            "title": "Periodo Fiscal"
          },
          "filas": {
            "items": {
              "$ref": "#/components/schemas/LibroVentasFila"
            },
            "type": "array",
            "title": "Filas"
          },
          "total_ventas_ves": {
            "type": "string",
            "title": "Total Ventas Ves"
          },
          "total_exento_ves": {
            "type": "string",
            "title": "Total Exento Ves"
          },
          "total_base_ves": {
            "type": "string",
            "title": "Total Base Ves"
          },
          "total_iva_ves": {
            "type": "string",
            "title": "Total Iva Ves"
          },
          "total_iva_retenido_ves": {
            "type": "string",
            "title": "Total Iva Retenido Ves"
          }
        },
        "type": "object",
        "required": [
          "periodo_fiscal",
          "filas",
          "total_ventas_ves",
          "total_exento_ves",
          "total_base_ves",
          "total_iva_ves",
          "total_iva_retenido_ves"
        ],
        "title": "LibroVentasResponse"
      },
...
    }
  }
}